重要改动！配置表读取操作修改 原先的读取操作为LuaTable，每次Get时会消耗很多性能，并且每Get一次都是一个新的对象 现在需要根据配置生成C#配置表代码，并且生成xluaBridge文件，每个配置表只需要Get1次，之后xlua会将lua配置数据拷贝到C#里

+++ CrazyGunplay/Assets/Scripts/Configs/Config_Bullet.cs
@@ -12,7 +12,7 @@ public interface Config_Bullet : IConfigBase
 
 	int bulletCount { get; set; }
 
-	List<float> bulletOffset { get; set; }
+	List<int> bulletOffset { get; set; }
 
 	float flySpeed { get; set; }
